+++ litho-widget/src/main/java/com/facebook/litho/widget/CardSpec.java
@@ -17,7 +17,7 @@
 package com.facebook.litho.widget;
 
 import static com.facebook.litho.widget.CardShadowDrawable.getShadowBottom;
-import static com.facebook.litho.widget.CardShadowDrawable.getShadowHorizontal;
+import static com.facebook.litho.widget.CardShadowDrawable.getShadowLeft;
 import static com.facebook.litho.widget.CardShadowDrawable.getShadowTop;
 import static com.facebook.yoga.YogaEdge.ALL;
 import static com.facebook.yoga.YogaEdge.BOTTOM;
@@ -100,10 +100,11 @@ class CardSpec {
       elevation = pixels(resources, DEFAULT_SHADOW_SIZE_DP);
     }
 
-    final int shadowTop = getShadowTop(elevation);
+    final float shadowDy = 0.5f * elevation;
+    final int shadowTop = getShadowTop(elevation, shadowDy);
     final int shadowBottom =
-        shadowBottomOverride == -1 ? getShadowBottom(elevation) : shadowBottomOverride;
-    final int shadowHorizontal = getShadowHorizontal(elevation);
+        shadowBottomOverride == -1 ? getShadowBottom(elevation, shadowDy) : shadowBottomOverride;
+    final int shadowHorizontal = getShadowLeft(elevation, 0f);
 
     return Column.create(c)
         .child(
@@ -131,6 +132,7 @@ class CardSpec {
                     .shadowEndColor(shadowEndColor)
                     .cornerRadiusPx(cornerRadius)
                     .shadowSizePx(elevation)
+                    .shadowDyPx(shadowDy)
                     .hideTopShadow(disableClipTopLeft && disableClipTopRight)
                     .hideBottomShadow(disableClipBottomLeft && disableClipBottomRight)
                     .positionType(ABSOLUTE)